+++ packages/sv-utils/src/tests/js/common/expression-from-string/run.ts
@@ -1,4 +1,5 @@
-import { common, exports, type AstTypes } from '../../../../tooling/js/index.ts';
+import type { AstTypes } from '../../../../tooling/index.ts';
+import { common, exports } from '../../../../tooling/js/index.ts';
 
 export function run(ast: AstTypes.Program): void {
 	exports.createDefault(ast, {