+++ packages/sv-utils/src/tests/js/exports/named-export-with-existing/run.ts
@@ -1,4 +1,5 @@
-import { common, variables, object, exports, type AstTypes } from '../../../../tooling/js/index.ts';
+import type { AstTypes } from '../../../../tooling/index.ts';
+import { common, variables, object, exports } from '../../../../tooling/js/index.ts';
 
 export function run(ast: AstTypes.Program): void {
 	const variableFallback = variables.declaration(ast, {