+++ packages/sv-utils/src/tests/js/object/ensure-nested-property/run.ts
@@ -1,4 +1,5 @@
-import { object, type AstTypes } from '../../../../tooling/js/index.ts';
+import type { AstTypes } from '../../../../tooling/index.ts';
+import { object } from '../../../../tooling/js/index.ts';
 import { getTestObjectExpression } from '../objectTestHelper.ts';
 
 export function run(ast: AstTypes.Program): void {