+++ packages/sv/src/cli/tests/snapshots/create-with-all-addons/src/lib/server/auth.ts
@@ -16,5 +16,7 @@ export const auth = betterAuth({
 			clientSecret: env.GITHUB_CLIENT_SECRET
 		}
 	},
-	plugins: [sveltekitCookies(getRequestEvent)] // make sure this is the last plugin in the array
+	plugins: [
+		sveltekitCookies(getRequestEvent) // make sure this is the last plugin in the array
+	]
 });